Move internal/cleanups to daemon/internal/cleanups
