+++ daemon/libnetwork/drivers/bridge/internal/nftabler/network.go
@@ -9,9 +9,9 @@ import (
 	"strings"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/cleanups"
 	"github.com/docker/docker/daemon/libnetwork/drivers/bridge/internal/firewaller"
 	"github.com/docker/docker/daemon/libnetwork/internal/nftables"
-	"github.com/docker/docker/internal/cleanups"
 	"go.opentelemetry.io/otel"
 )
